requirements.txt
GitPython>=3.1.24
PyGithub>=1.55